project/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.44
jinja2==3.1.2
python-decouple==3.8
bcrypt==4.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
starlette<0.27.0  # Fix conflict
